Fixed the Payload file for the 5-second transmissions only, nothing done with confirmation.
The ground worked for awhile, then I destroyed the working copy by writing over it with a bad one. But I did find the issue that caused the YYYYYYYY's to appear at the output. Description are in comments inside the file. Finding what caused the YYYYYYYYYYYYY's  took 3 hours, wayyyyyyyy too long.